.r Move managed-root staging helpers to staging.go (go test green; automated symbol move)
internal/authpolicy/manage.go was 2073 lines spanning dispatch,
parsing, commands, status, bootstrap, validators, file staging,
policy mutation, and rendering — all in one file.  Extract the
managed-root file-ops block (14 functions: canonicalDestinationPath,
canonicalDestinationPathPart, ensureDirectory, validateManagedPath,
writeManagedRootMarker, isWorkcellManagedRoot, cleanupStagedFile,
stageExistingFile, restoreStagedFile, openManagedRoot,
managedRelativePath, writeSourceFile, normalizeRootRelativePath,
reserveRootTempPath, randomTempSuffix) into a new staging.go.

manage.go drops 4 imports (crypto/rand, errors, rootio, secretfile)
that are now only used by staging.go.  Zero behavior change; same
package, all symbols stay package-internal.
