+++ internal/authpolicy/manage.go
@@ -4,8 +4,6 @@
 package authpolicy
 
 import (
-	"crypto/rand"
-	"errors"
 	"flag"
 	"fmt"
 	"io"
@@ -18,8 +16,6 @@ import (
 	"github.com/omkhar/workcell/internal/cliexit"
 	"github.com/omkhar/workcell/internal/pathutil"
 	"github.com/omkhar/workcell/internal/providerid"
-	"github.com/omkhar/workcell/internal/rootio"
-	"github.com/omkhar/workcell/internal/secretfile"
 )
 
 var (
@@ -1084,268 +1080,10 @@ func credentialStateIsReady(state string) bool {
 	}
 }
 
-func summarizeBootstrap(agent string, selected map[string]any, inputKinds, resolvers, resolutionStates, providerReadyStates map[string]string) bootstrapSummary {
-	switch agent {
-	case providerid.Codex:
-		if readiness, ok := providerReadyStates["codex_auth"]; ok {
-			if credentialStateIsReady(readiness) {
-				if inputKinds["codex_auth"] == "resolver" {
-					return bootstrapSummary{
-						state:    "ready",
-						path:     "host-resolver",
-						support:  "repo-required",
-						handoff:  "none",
-						doc:      "docs/examples/quickstart-codex.md",
-						nextStep: "none",
-					}
-				}
-				return bootstrapSummary{
-					state:    "ready",
-					path:     "direct-staged",
-					support:  "repo-required",
-					handoff:  "none",
-					doc:      "docs/examples/quickstart-codex.md",
-					nextStep: "none",
-				}
-			}
-			if readiness == "configured-only" && resolvers["codex_auth"] == "codex-home-auth-file" {
-				return bootstrapSummary{
-					state:    "configured-only",
-					path:     "host-resolver",
-					support:  "repo-required",
-					handoff:  "host-provider-cache",
-					doc:      "docs/examples/quickstart-codex.md",
-					nextStep: "stage-reviewed-codex-auth",
-				}
-			}
-		}
-		return defaultBootstrapSummary(agent)
-	case providerid.Claude:
-		for _, key := range []string{"claude_api_key", "claude_auth"} {
-			if credentialStateIsReady(providerReadyStates[key]) {
-				return bootstrapSummary{
-					state:    "ready",
-					path:     "direct-staged",
-					support:  "repo-required",
-					handoff:  "none",
-					doc:      "docs/examples/quickstart-claude.md",
-					nextStep: "none",
-				}
-			}
-		}
-		if resolutionStates["claude_auth"] == "configured-only" && resolvers["claude_auth"] == "claude-macos-keychain" {
-			return bootstrapSummary{
-				state:    "configured-only",
-				path:     "host-export-scaffold",
-				support:  "manual",
-				handoff:  "host-export",
-				doc:      "docs/examples/quickstart-claude.md",
-				nextStep: "stage-reviewed-claude-auth-or-api-key",
-			}
-		}
-		return defaultBootstrapSummary(agent)
-	case providerid.Gemini:
-		for _, key := range []string{"gemini_env", "gemini_oauth"} {
-			if credentialStateIsReady(providerReadyStates[key]) {
-				return bootstrapSummary{
-					state:    "ready",
-					path:     "direct-staged",
-					support:  "repo-required",
-					handoff:  "none",
-					doc:      "docs/examples/quickstart-gemini.md",
-					nextStep: "none",
-				}
-			}
-		}
-		if credentialStateIsReady(providerReadyStates["gemini_projects"]) {
-			return bootstrapSummary{
-				state:    "supplemental-only",
-				path:     "project-registry-supplement",
-				support:  "manual",
-				handoff:  "host-stage-file",
-				doc:      "docs/examples/quickstart-gemini.md",
-				nextStep: "stage-reviewed-gemini-env-or-oauth",
-			}
-		}
-		if credentialStateIsReady(providerReadyStates["gcloud_adc"]) {
-			return bootstrapSummary{
-				state:    "supplemental-only",
-				path:     "vertex-supplement",
-				support:  "manual",
-				handoff:  "host-stage-file",
-				doc:      "docs/examples/gemini-vertex-setup.md",
-				nextStep: "stage-reviewed-gemini-env-or-oauth",
-			}
-		}
-		return defaultBootstrapSummary(agent)
-	default:
-		return bootstrapSummary{}
-	}
-}
-
-func defaultBootstrapSummary(agent string) bootstrapSummary {
-	switch agent {
-	case providerid.Codex:
-		return bootstrapSummary{
-			state:    "not-configured",
-			path:     "direct-staged",
-			support:  "repo-required",
-			handoff:  "host-stage-file",
-			doc:      "docs/examples/quickstart-codex.md",
-			nextStep: "stage-reviewed-codex-auth",
-		}
-	case providerid.Claude:
-		return bootstrapSummary{
-			state:    "not-configured",
-			path:     "direct-staged",
-			support:  "repo-required",
-			handoff:  "host-stage-file",
-			doc:      "docs/examples/quickstart-claude.md",
-			nextStep: "stage-reviewed-claude-auth-or-api-key",
-		}
-	case providerid.Gemini:
-		return bootstrapSummary{
-			state:    "not-configured",
-			path:     "direct-staged",
-			support:  "repo-required",
-			handoff:  "host-stage-file",
-			doc:      "docs/examples/quickstart-gemini.md",
-			nextStep: "stage-reviewed-gemini-env-or-oauth",
-		}
-	default:
-		return bootstrapSummary{}
-	}
-}
-
-func bootstrapSummaryForCredential(agent, credential string, report credentialSelectionReport) bootstrapSummary {
-	switch credential {
-	case "codex_auth":
-		if report.inputKind == "resolver" {
-			if credentialStateIsReady(report.readiness) {
-				return bootstrapSummary{
-					state:    "ready",
-					path:     "host-resolver",
-					support:  "repo-required",
-					handoff:  "none",
-					doc:      "docs/examples/quickstart-codex.md",
-					nextStep: "none",
-				}
-			}
-			return bootstrapSummary{
-				state:    report.readiness,
-				path:     "host-resolver",
-				support:  "repo-required",
-				handoff:  "host-provider-cache",
-				doc:      "docs/examples/quickstart-codex.md",
-				nextStep: "stage-reviewed-codex-auth",
-			}
-		}
-		return bootstrapSummary{
-			state:    report.readiness,
-			path:     "direct-staged",
-			support:  "repo-required",
-			handoff:  bootstrapHandoffForReadiness(report.readiness),
-			doc:      "docs/examples/quickstart-codex.md",
-			nextStep: bootstrapNextStepForReadiness(report.readiness, "stage-reviewed-codex-auth"),
-		}
-	case "claude_auth":
-		if report.inputKind == "resolver" {
-			if credentialStateIsReady(report.readiness) {
-				return bootstrapSummary{
-					state:    "ready",
-					path:     "host-export-scaffold",
-					support:  "manual",
-					handoff:  "none",
-					doc:      "docs/examples/quickstart-claude.md",
-					nextStep: "none",
-				}
-			}
-			return bootstrapSummary{
-				state:    report.readiness,
-				path:     "host-export-scaffold",
-				support:  "manual",
-				handoff:  "host-export",
-				doc:      "docs/examples/quickstart-claude.md",
-				nextStep: "stage-reviewed-claude-auth-or-api-key",
-			}
-		}
-		return bootstrapSummary{
-			state:    report.readiness,
-			path:     "direct-staged",
-			support:  "repo-required",
-			handoff:  bootstrapHandoffForReadiness(report.readiness),
-			doc:      "docs/examples/quickstart-claude.md",
-			nextStep: bootstrapNextStepForReadiness(report.readiness, "stage-reviewed-claude-auth-or-api-key"),
-		}
-	case "claude_api_key", "claude_mcp":
-		return bootstrapSummary{
-			state:    report.readiness,
-			path:     "direct-staged",
-			support:  "repo-required",
-			handoff:  bootstrapHandoffForReadiness(report.readiness),
-			doc:      "docs/examples/quickstart-claude.md",
-			nextStep: bootstrapNextStepForReadiness(report.readiness, "stage-reviewed-claude-auth-or-api-key"),
-		}
-	case "gemini_env", "gemini_oauth", "gemini_projects":
-		return bootstrapSummary{
-			state:    report.readiness,
-			path:     "direct-staged",
-			support:  "repo-required",
-			handoff:  bootstrapHandoffForReadiness(report.readiness),
-			doc:      "docs/examples/quickstart-gemini.md",
-			nextStep: bootstrapNextStepForReadiness(report.readiness, "stage-reviewed-gemini-env-or-oauth"),
-		}
-	case "gcloud_adc":
-		return bootstrapSummary{
-			state:    report.readiness,
-			path:     "vertex-supplement",
-			support:  "manual",
-			handoff:  bootstrapHandoffForReadiness(report.readiness),
-			doc:      "docs/examples/gemini-vertex-setup.md",
-			nextStep: bootstrapNextStepForReadiness(report.readiness, "stage-reviewed-gemini-env-or-oauth"),
-		}
-	default:
-		return defaultBootstrapSummary(agent)
-	}
-}
-
-func bootstrapHandoffForReadiness(readiness string) string {
-	if credentialStateIsReady(readiness) {
-		return "none"
-	}
-	return "host-stage-file"
-}
-
-func bootstrapNextStepForReadiness(readiness, nextStep string) string {
-	if credentialStateIsReady(readiness) {
-		return "none"
-	}
-	return nextStep
-}
-
-func printBootstrapSummary(stdout io.Writer, summary bootstrapSummary) {
-	if summary.path == "" {
-		return
-	}
-	fmt.Fprintln(stdout, "provider_bootstrap_state="+summary.state)
-	fmt.Fprintln(stdout, "provider_bootstrap_path="+summary.path)
-	fmt.Fprintln(stdout, "provider_bootstrap_support="+summary.support)
-	fmt.Fprintln(stdout, "provider_bootstrap_handoff="+summary.handoff)
-	fmt.Fprintln(stdout, "provider_bootstrap_doc="+summary.doc)
-	fmt.Fprintln(stdout, "provider_bootstrap_next_step="+summary.nextStep)
-}
-
-func printCredentialBootstrapSummary(stdout io.Writer, summary bootstrapSummary) {
-	if summary.path == "" {
-		return
-	}
-	fmt.Fprintln(stdout, "bootstrap_state="+summary.state)
-	fmt.Fprintln(stdout, "bootstrap_path="+summary.path)
-	fmt.Fprintln(stdout, "bootstrap_support="+summary.support)
-	fmt.Fprintln(stdout, "bootstrap_handoff="+summary.handoff)
-	fmt.Fprintln(stdout, "bootstrap_doc="+summary.doc)
-	fmt.Fprintln(stdout, "bootstrap_next_step="+summary.nextStep)
-}
+// Bootstrap-summary helpers (summarizeBootstrap, defaultBootstrapSummary,
+// bootstrapSummaryForCredential, bootstrapHandoffForReadiness,
+// bootstrapNextStepForReadiness, printBootstrapSummary,
+// printCredentialBootstrapSummary) live in bootstrap.go.
 
 func boolToInt(value bool) int {
 	if value {
@@ -1406,172 +1144,12 @@ func desiredSelectorsFromPolicy(policyPath string, credential string, agent stri
 	return desiredSelectors(credentials, credential, agent)
 }
 
-func canonicalDestinationPath(managedRoot string, credential string) (string, error) {
-	parts, ok := canonicalCredentialDestinations[credential]
-	if !ok {
-		return "", die(fmt.Sprintf("workcell auth set does not manage %s automatically", credential))
-	}
-	return filepath.Join(managedRoot, parts[0], parts[1]), nil
-}
-
-func canonicalDestinationPathPart(credential string) string {
-	parts := canonicalCredentialDestinations[credential]
-	return filepath.Join(parts[0], parts[1])
-}
-
-func ensureDirectory(path string) error {
-	if err := os.MkdirAll(path, 0o700); err != nil {
-		return err
-	}
-	return os.Chmod(path, 0o700)
-}
-
-func validateManagedPath(managedRoot string, path string, label string) error {
-	if err := requireNoSymlinkInPathChain(path, label); err != nil {
-		return err
-	}
-	return requirePathWithin(managedRoot, path, label)
-}
-
-func writeManagedRootMarker(managedRoot string) error {
-	if err := ensureDirectory(managedRoot); err != nil {
-		return err
-	}
-	managedRootFS, err := os.OpenRoot(managedRoot)
-	if err != nil {
-		return err
-	}
-	defer managedRootFS.Close()
-	return rootio.WriteFileAtomic(managedRootFS, managedRootMarker, []byte("managed_by=workcell\n"), 0o600, "."+managedRootMarker+"-")
-}
-
-func isWorkcellManagedRoot(path string) bool {
-	_, err := os.Stat(filepath.Join(path, managedRootMarker))
-	return err == nil
-}
-
-func cleanupStagedFile(root *os.Root, path string) error {
-	if path == "" {
-		return nil
-	}
-	if err := root.Remove(path); err != nil && !os.IsNotExist(err) {
-		return err
-	}
-	return nil
-}
-
-func stageExistingFile(root *os.Root, path string) (string, error) {
-	if path == "" {
-		return "", nil
-	}
-	cleanPath, err := normalizeRootRelativePath(path)
-	if err != nil {
-		return "", err
-	}
-	if _, err := root.Stat(cleanPath); err != nil {
-		if os.IsNotExist(err) {
-			return "", nil
-		}
-		return "", err
-	}
-	tempPath, err := reserveRootTempPath(root, filepath.Dir(cleanPath), ".workcell-stage-")
-	if err != nil {
-		return "", err
-	}
-	if err := root.Rename(cleanPath, tempPath); err != nil {
-		return "", err
-	}
-	return tempPath, nil
-}
-
-func restoreStagedFile(root *os.Root, stagedPath string, destination string) error {
-	if stagedPath == "" || destination == "" {
-		return nil
-	}
-	cleanDestination, err := normalizeRootRelativePath(destination)
-	if err != nil {
-		return err
-	}
-	if err := root.Rename(stagedPath, cleanDestination); err != nil {
-		return err
-	}
-	return nil
-}
-
-func openManagedRoot(managedRoot string) (*os.Root, error) {
-	if err := ensureDirectory(managedRoot); err != nil {
-		return nil, err
-	}
-	return os.OpenRoot(managedRoot)
-}
-
-func managedRelativePath(managedRoot string, path string, label string) (string, error) {
-	if path == "" {
-		return "", nil
-	}
-	return rootio.RelativePathWithin(managedRoot, path, label)
-}
-
-func writeSourceFile(root *os.Root, source string, destination string) error {
-	in, err := secretfile.Open(source, "credential source", os.Getuid())
-	if err != nil {
-		return err
-	}
-	defer in.Close()
-	return rootio.WriteFileAtomicFromReader(root, destination, in, 0o600, ".workcell-auth-")
-}
-
-func normalizeRootRelativePath(path string) (string, error) {
-	if filepath.IsAbs(path) {
-		return "", fmt.Errorf("path must be relative to managed root: %s", path)
-	}
-	cleanPath := filepath.Clean(path)
-	if cleanPath == "." || cleanPath == string(filepath.Separator) {
-		return "", fmt.Errorf("path must name a file within the managed root: %s", path)
-	}
-	return cleanPath, nil
-}
-
-func reserveRootTempPath(root *os.Root, parent string, prefix string) (string, error) {
-	parent = filepath.Clean(parent)
-	if parent == "." {
-		parent = ""
-	}
-	for attempt := 0; attempt < 32; attempt++ {
-		suffix, err := randomTempSuffix()
-		if err != nil {
-			return "", err
-		}
-		name := prefix + suffix + ".tmp"
-		tempPath := name
-		if parent != "" {
-			tempPath = filepath.Join(parent, name)
-		}
-		tempFile, err := root.OpenFile(tempPath, os.O_CREATE|os.O_EXCL|os.O_WRONLY, 0o600)
-		if err == nil {
-			if closeErr := tempFile.Close(); closeErr != nil {
-				_ = root.Remove(tempPath)
-				return "", closeErr
-			}
-			if removeErr := root.Remove(tempPath); removeErr != nil {
-				return "", removeErr
-			}
-			return tempPath, nil
-		}
-		if !errors.Is(err, os.ErrExist) {
-			return "", err
-		}
-	}
-	return "", fmt.Errorf("unable to allocate temporary staging path under %s", root.Name())
-}
-
-func randomTempSuffix() (string, error) {
-	var data [8]byte
-	if _, err := rand.Read(data[:]); err != nil {
-		return "", err
-	}
-	return fmt.Sprintf("%x", data[:]), nil
-}
+// Managed-root file ops (canonicalDestinationPath, ensureDirectory,
+// validateManagedPath, writeManagedRootMarker, isWorkcellManagedRoot,
+// cleanupStagedFile, stageExistingFile, restoreStagedFile,
+// openManagedRoot, managedRelativePath, writeSourceFile,
+// normalizeRootRelativePath, reserveRootTempPath, randomTempSuffix)
+// live in staging.go.
 
 func loadMutablePolicy(policyPath string) (map[string]any, error) {
 	policy, err := loadRawPolicy(policyPath)
